+++ src/lib/actions/onacceptEventlistener.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function(settings) {
+  window.addEventListener('CookiebotOnAccept', function(e){
+    CookieHelper.selection();
+    });    
+};