+++ src/lib/actions/ondeclineEventlistener.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function(settings) {
+  window.addEventListener('CookiebotOnDecline', function(e){
+    CookieHelper.decline();
+    });  
+};